examples/rag-pinecone-research-surya/requirements.txt
langchain-community
langchain-pinecone
pinecone-client
openai
judgeval
tqdm
python-dotenv
pypdf